Fix #6 by properly inheriting CALayer's shadowOpacity type, from CGFloat to float.

+++ Classes/YIInnerShadowLayer.m
@@ -86,7 +86,7 @@
     [self setNeedsLayout];
 }
 
-- (void)setShadowOpacity:(CGFloat)shadowOpacity
+- (void)setShadowOpacity:(float)shadowOpacity
 {
     [super setShadowOpacity:shadowOpacity];
     [self setNeedsLayout];
